chore: Update npm dependencies and remove unused router
This commit updates the npm dependencies in the project, adding the `axios` package with version `1.7.2` and the `timeago.js` package with version `4.0.2`. It also removes the unused router file `src/router/index.js`. These changes ensure that the project is using the latest stable versions of the dependencies and removes unnecessary code.