chore: Update scrollbind function in HomeView and ProfileView
This commit updates the scrollbind function in the HomeView and ProfileView components. It adds a condition to check if the length of the data is less than 9, and if so, sets the showloadmore value to false. This ensures that the "Load More" button is not displayed when there are fewer than 9 items in the data.