refactor: Optimize image loading in HomeView.vue
This commit optimizes the image loading logic in the HomeView.vue component. It simplifies the code by removing unnecessary div elements and improves the rendering of images. The changes also include fixing a bug where images were not displayed correctly when the attachment type was 'image'. These improvements enhance the page load performance and provide a better user experience.